+++ src/03-bot/NasdaqBot/Models/StockOrderRequest.cs
@@ -4,6 +4,6 @@ namespace NasdaqBot.Models
     {
         public int Amount { get; set; }
         public string StockSymbol { get; set; }
-        public double OrderLimit { get; set; }
+        public decimal OrderLimit { get; set; }
     }
 }
